+++ backend/src/main.py
@@ -4,14 +4,9 @@
 from __future__ import annotations
 
 import logging
-import sys
-from datetime import datetime
-from logging import Handler, StreamHandler
-from pathlib import Path
 from signal import SIGTERM
-from typing import TYPE_CHECKING, Final, LiteralString
+from typing import TYPE_CHECKING, Final
 
-import colorlog
 import core
 import game
 import gui
@@ -135,41 +130,10 @@ if __name__ == '__main__':
     app.extra['host'] = cli.args.host
     app.extra['port'] = cli.args.port
 
-    def _get_log_format(*, use_colors: bool = False) -> str:
-        time: LiteralString = '%(asctime)s'
-        level: LiteralString = '%(levelname)s'
-        if use_colors:
-            time = f'%(light_black)s{time}%(reset)s'
-            level = f'%(bold)s%(log_color)s{level}%(reset)s'
-        return f'{time} {level} %(message)s'
-
     # Configure logging
-    datefmt: LiteralString = '%H:%M:%S'
-    log_dir: Final[Path] = Path(LOG_DIR_NAME)
-    if not log_dir.exists():
-        log_dir.mkdir()
-    file: Path = log_dir / Path(f'{datetime.now().strftime("%Y-%m-%d")}.log')
-    logging_handlers: list[Handler] = [logging.FileHandler(file, mode='a')]
-    if not cli.args.silent:
-        console_logger: StreamHandler = logging.StreamHandler(sys.stdout)
-        console_logger.formatter = colorlog.ColoredFormatter(
-            fmt=_get_log_format(use_colors=True),
-            datefmt=datefmt,
-            log_colors={
-                'DEBUG': 'cyan',
-                'INFO': 'green',
-                'WARNING': 'yellow',
-                'ERROR': 'red',
-                'CRITICAL': 'red,bg_white',
-            },
-        )
-        logging_handlers.append(console_logger)
-    logging.basicConfig(
-        level=logging.DEBUG if app.debug else logging.INFO,
-        format=_get_log_format(use_colors=False),
-        datefmt=datefmt,
-        handlers=logging_handlers,
-    )
+    silent_logging: bool = cli.args.silent
+    log_level: int = logging.DEBUG if app.debug else logging.INFO
+    core.configure_logging(LOG_DIR_NAME, level=log_level, silent=silent_logging)
 
     # Check for new releases in the Github releases page
     if cli.args.check_for_releases:
